+++ NEWS.md
@@ -3,6 +3,10 @@
 * Fix a bug in constructing empty blocks for extending
 * You can now pass in additional arguments to the OpenAPI doc UI with
   `api_doc_setting()`
+* More powerful report support. Added `api_report()` as a parallel to `@report``
+  but with even more settings. Parameters are now type checked using the same
+  facilities as the request handlers, and much more comprehensive OpenAPI
+  documentation is provided.
 
 # plumber2 0.1.0